Get the TransferMatrix working by dropping the Clebsch-Gordan work
Expanding the beam in terms of a small number of spherical harmonics
means that the beam is actually sensitive to the entire sky (at a small
level). This is unacceptable, so I'd rather keep the full description of
the beam and use alm2map/map2alm to multiply it with the fringe pattern
instead of using the Clebsch-Gordan coefficients.

Essentially this reverts a failed experiment.


Former-commit-id: 9ffedc5d6a0de9639a7cbdca1878cfe535897c36